Forward ports directly in the docker-compose file for compatibility. We aren't making use of the differences in  devcontainer 'forwardPorts (#96)
We pass several ports from this devcontainer compose stack (NATS, Hydra, CRDB...) through to the host using the ports option directly in the docker-compose.yml file... but we have been omitting the port used by the governor-api app itself from the compose file and instead forwarded it using the devcontainer.json forwardPorts definition.

This works seamlessly in VSCode, however, forwardPorts is unsupported by the devcontainer CLI for anyone using another editor/IDE. The main difference between forwarding a port in the compose file vs. using forwardPorts is that forwardPorts is necessary for services that only listen on 127.0.0.1 - however, this service listens on 0.0.0.0, thus there is no functional difference here, so proposing this change for QoL/compatibility reasons.