Membership approver groups (#101)
Add approver group field to groups, allowing for the designation of group membership request approval to users that are not a member of the group itself. The use case is that a certain group may confer a high level of access to certain systems, which no users should have on a standing basis. This allows such a high-privileged group to remain 'normally empty', while still making it useful for break-glass access situations.